face_recognition_toGCP/requirements.txt
Cmake==3.28.1
imutils==0.5.4
numpy==1.19.5
opencv-python==4.5.4.60
Pillow==8.3.1
face-recognition==1.3.0
face-recognition-models==0.3.0
flask==3.0.1
google-cloud-storage==2.14.0
